// src/app/Model/production.ts
export class LineCode {
  ProductionLineCode: string = "";
  ProductiongLineName: string = "";
}
// ...
export class ProductionOrder {
  PartNo: string = "";
  ProductionOrderNo: string = "";
  OrderQTY: number = 0;
  FindQTY: number = 0;
}

export class CusOrder {
  PartNo: string = "";
  CusOrderNo: string = "";
  OrderQTY: number = 0;
  FindQTY: number = 0;
  PONum: string = '';
  FinDate: string = '';
  ShipLocation: string = '';
  TD001: string = '';
  TD002: string = '';
  D003: string = '';
}
// ...
export class OrderStatusInfo {
  ProductionHead: string = '';
  ProductionOrder: string = '';
  ProductionOrderStatus: string = '';
  SchedulingDate: string = '';
  OrderDate: string = '';
  OrderStatus: string = '';
  OrderHeadCode:string;
  Salesperson: string = '';
  ETD: string = '';
  BillingDate: string = '';
  CustomerName: string = '';
  CusOrder: string = '';
  ItemCode: string = '';
  ItemName: string = '';
  Quantity: number = 0;
  Weight: number = 0;
  MaterialType: string = '';
  Flask: string = '';
  AsemblingLineCode: string = '';
  AsemblingLineName: string = '';
  Details: Array<OrderStatusInfoDetail> = new Array<OrderStatusInfoDetail>();
  OrdererrStatus: string = '';
  Remark: string = '';
  WorkOrderStatus: boolean;
  SalesOrderStatus:string;
  LateStart:boolean = false;
  MailFlag:boolean = false;
  light:string = '';
}

export class OrderStatusInfoDetail {
  ProductionOrderCode: string = '';
  ProductionHead: string = '';
  ProductionCode: string = '';
  ProductionOrderStatus: string = '';
  MainProductionHead: string = '';
  MainProductionCode: string = '';
  ProductionQuantity: string = '';
  ProcessSequence: string = '';
  WIPProcessCode: string = '';
  LineCode: string = '';
  LineName: string = '';
  PlanStartDate: string = '';
  PlanFinishDate: string = '';
  FinishQuantity: number = 0;
  UserCode: string = '';
  UserName: string = '';
  StartDate: string = '';
  FinishDate: string = '';
  StartTime: string = '';
  EndTime: string = '';
}